requirements.txt
anthropic
httpx
python-dotenv
pypdf
sentence-transformers==2.7.0
faiss-cpu
streamlit
fpdf2
